resize columns, new two columns in productTradeName, multiselect in grid with metakey

+++ src/app/modules/universal-components/components/table/table.service.ts
@@ -35,8 +35,8 @@ export class TableService {
         return 'boolean';
       case 'Int32':
         return 'numeric';
-      case 'Int32':
-        return 'numeric';
+      case 'Single':
+        return 'float';
       case 'String':
         return 'text';
       case 'DateTime':
